update fakefullscreen with the ideas from the lockfullscreen commit
this makes it so that when in fullscreen mode you can't navigate away from
a fullscreen window using focusstack(). this behaviour didn't make sense
since, if you navigated away, you couldn't actually tell which window was
focused and this made it annoying to un-fullscreen. now you can navigate
away from a fakefullscreen window but not a real fullscreen window.